Extract parsing of command line args
