+++ src/HtmlGenerator/Program.cs
@@ -17,201 +17,17 @@ namespace Microsoft.SourceBrowser.HtmlGenerator
     {
         private static void Main(string[] args)
         {
-            if (args.Length == 0)
+            var options = CommandLineOptions.Parse(args);
+
+            if (options.Properties.Count == 0)
             {
                 PrintUsage();
                 return;
             }
 
-            var projects = new List<string>();
-            var properties = new Dictionary<string, string>();
-            var emitAssemblyList = false;
-            var doNotIncludeReferencedProjects = false;
-            var force = false;
-            var noBuiltInFederations = false;
-            var offlineFederations = new Dictionary<string, string>();
-            var federations = new HashSet<string>();
-            var serverPathMappings = new Dictionary<string, string>();
-            var pluginBlacklist = new List<string>();
-            var rootPath = (string)null;
-
-            foreach (var arg in args)
-            {
-                if (arg.StartsWith("/out:"))
-                {
-                    Paths.SolutionDestinationFolder = Path.GetFullPath(arg.Substring("/out:".Length).StripQuotes());
-                    continue;
-                }
-
-                if (arg.StartsWith("/serverPath:"))
-                {
-                    // Allowed forms:
-                    // /serverPath:a=b
-                    // /serverPath:"a=b" (for backwards compatibility)
-                    // /serverPath:"a=1"="b=2"
-                    // /serverPath:"a"="b" (for consistency to make it easy to produce a safe form)
-
-                    var match = Regex.Match(
-                        arg.Substring("/serverPath:".Length),
-                        @"\A(?:
-                            # Each side may be quoted or unquoted but may only contain '=' if quoted
-                            (?:(?<from>[^""=]*)|""(?<from>[^""]*)"")=(?:(?<to>[^""=]*)|""(?<to>[^""]*)"")
-                            |
-                            # Backwards compatibility, not advertised as an option because it doesn't allow '=' even though there are quotes
-                            ""(?<from>[^""=]*)=(?<to>[^""=]*)""
-                        )\Z", RegexOptions.IgnorePatternWhitespace);
-
-                    if (!match.Success)
-                    {
-                        Log.Write("Server path argument usage: /serverPath:\"path to local repository root\"=\"root URL\"" + Environment.NewLine +
-                                  "Quotes are optional if you have no spaces or equals signs but recommended. Paths relative to the local repository root will be appended to the root URL.", ConsoleColor.Red);
-                        continue;
-                    }
-
-                    serverPathMappings.Add(Path.GetFullPath(match.Groups["from"].Value), match.Groups["to"].Value);
-                    continue;
-                }
-
-                if (arg == "/force")
-                {
-                    force = true;
-                    continue;
-                }
-
-                if (arg.StartsWith("/in:"))
-                {
-                    string inputPath = arg.Substring("/in:".Length).StripQuotes();
-                    try
-                    {
-                        if (!File.Exists(inputPath))
-                        {
-                            continue;
-                        }
-
-                        string[] paths = File.ReadAllLines(inputPath);
-                        foreach (string path in paths)
-                        {
-                            AddProject(projects, path);
-                        }
-                    }
-                    catch
-                    {
-                        Log.Write("Invalid argument: " + arg, ConsoleColor.Red);
-                    }
-
-                    continue;
-                }
-
-                if (arg.StartsWith("/p:"))
-                {
-                    var match = Regex.Match(arg, "/p:(?<name>[^=]+)=(?<value>.+)");
-                    if (match.Success)
-                    {
-                        var propertyName = match.Groups["name"].Value;
-                        var propertyValue = match.Groups["value"].Value;
-                        properties.Add(propertyName, propertyValue);
-                        continue;
-                    }
-                }
-
-                if (arg == "/assemblylist")
-                {
-                    emitAssemblyList = true;
-                    continue;
-                }
-
-                if (string.Equals(arg, "/donotincludereferencedprojects", StringComparison.OrdinalIgnoreCase))
-                {
-                    doNotIncludeReferencedProjects = true;
-                    continue;
-                }
-
-                if (arg == "/nobuiltinfederations")
-                {
-                    noBuiltInFederations = true;
-                    Log.Message("Disabling built-in federations.");
-                    continue;
-                }
-
-                if (arg.StartsWith("/federation:"))
-                {
-                    string server = arg.Substring("/federation:".Length);
-                    Log.Message($"Adding federation '{server}'.");
-                    federations.Add(server);
-                    continue;
-                }
-
-                if (arg.StartsWith("/offlinefederation:"))
-                {
-                    var match = Regex.Match(arg, "/offlinefederation:(?<server>[^=]+)=(?<file>.+)");
-                    if (match.Success)
-                    {
-                        var server = match.Groups["server"].Value;
-                        var assemblyListFileName = match.Groups["file"].Value;
-                        offlineFederations[server] = assemblyListFileName;
-                        Log.Message($"Adding federation '{server}' (offline from '{assemblyListFileName}').");
-                        continue;
-                    }
-                    continue;
-                }
-
-                if (string.Equals(arg, "/noplugins", StringComparison.OrdinalIgnoreCase))
-                {
-                    SolutionGenerator.LoadPlugins = false;
-                    continue;
-                }
-
-                if (string.Equals(arg, "/useplugins", StringComparison.OrdinalIgnoreCase))
-                {
-                    SolutionGenerator.LoadPlugins = true;
-                    continue;
-                }
-
-                if (arg.StartsWith("/noplugin:"))
-                {
-                    pluginBlacklist.Add(arg.Substring("/noplugin:".Length));
-                    continue;
-                }
-
-                if (arg == "/excludetests")
-                {
-                    SolutionGenerator.ExcludeTests = true;
-                    continue;
-                }
-
-                if (arg.StartsWith("/root:", StringComparison.Ordinal))
-                {
-                    rootPath = Path.GetFullPath(arg.Substring("/root:".Length).StripQuotes());
-                    continue;
-                }
-
-                try
-                {
-                    AddProject(projects, arg);
-                }
-                catch (Exception ex)
-                {
-                    Log.Write("Exception: " + ex.ToString(), ConsoleColor.Red);
-                }
-            }
-
-            if (projects.Count == 0)
-            {
-                PrintUsage();
-                return;
-            }
-
-            if (rootPath is object)
-            {
-                foreach (var project in projects)
-                {
-                    if (!Paths.IsOrContains(rootPath, project))
-                    {
-                        Log.Exception("If /root is specified, it must be an ancestor folder of all specified projects.", isSevere: true);
-                        return;
-                    }
-                }
-            }
+            Paths.SolutionDestinationFolder = options.SolutionDestinationFolder;
+            SolutionGenerator.LoadPlugins = options.LoadPlugins;
+            SolutionGenerator.ExcludeTests = options.ExcludeTests;
 
             AssertTraceListener.Register();
             AppDomain.CurrentDomain.FirstChanceException += FirstChanceExceptionHandler.HandleFirstChanceException;
@@ -228,7 +44,7 @@ namespace Microsoft.SourceBrowser.HtmlGenerator
             var websiteDestination = Paths.SolutionDestinationFolder;
 
             // Warning, this will delete and recreate your destination folder
-            Paths.PrepareDestinationFolder(force);
+            Paths.PrepareDestinationFolder(options.Force);
 
             Paths.SolutionDestinationFolder = Path.Combine(Paths.SolutionDestinationFolder, "index"); //The actual index files need to be written to the "index" subdirectory
 
@@ -241,52 +57,23 @@ namespace Microsoft.SourceBrowser.HtmlGenerator
             {
                 var federation = new Federation();
 
-                if (!noBuiltInFederations)
+                if (!options.NoBuiltInFederations)
                 {
                     federation.AddFederations(Federation.DefaultFederatedIndexUrls);
                 }
 
-                federation.AddFederations(federations);
+                federation.AddFederations(options.Federations);
 
-                foreach (var entry in offlineFederations)
+                foreach (var entry in options.OfflineFederations)
                 {
                     federation.AddFederation(entry.Key, entry.Value);
                 }
 
-                IndexSolutions(projects, properties, federation, serverPathMappings, pluginBlacklist, doNotIncludeReferencedProjects, rootPath);
-                FinalizeProjects(emitAssemblyList, federation);
-                WebsiteFinalizer.Finalize(websiteDestination, emitAssemblyList, federation);
+                IndexSolutions(options.Projects, options.Properties, federation, options.ServerPathMappings, options.PluginBlacklist, options.DoNotIncludeReferencedProjects, options.RootPath);
+                FinalizeProjects(options.EmitAssemblyList, federation);
+                WebsiteFinalizer.Finalize(websiteDestination, options.EmitAssemblyList, federation);
             }
             Log.Close();
-        }
-
-        private static void AddProject(List<string> projects, string path)
-        {
-            var project = Path.GetFullPath(path);
-            if (IsSupportedProject(project))
-            {
-                projects.Add(project);
-            }
-            else
-            {
-                Log.Exception("Project not found or not supported: " + path, isSevere: false);
-            }
-        }
-
-        private static bool IsSupportedProject(string filePath)
-        {
-            if (!File.Exists(filePath))
-            {
-                return false;
-            }
-
-            return filePath.EndsWith(".sln", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".binlog", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".buildlog", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".csproj", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".vbproj", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".dll", StringComparison.OrdinalIgnoreCase) ||
-                   filePath.EndsWith(".exe", StringComparison.OrdinalIgnoreCase);
         }
 
         private static void PrintUsage()
@@ -324,9 +111,9 @@ namespace Microsoft.SourceBrowser.HtmlGenerator
 
         private static void IndexSolutions(
             IEnumerable<string> solutionFilePaths,
-            Dictionary<string, string> properties,
+            IReadOnlyDictionary<string, string> properties,
             Federation federation,
-            Dictionary<string, string> serverPathMappings,
+            IReadOnlyDictionary<string, string> serverPathMappings,
             IEnumerable<string> pluginBlacklist,
             bool doNotIncludeReferencedProjects = false,
             string rootPath = null)
